requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alpaca-trade-api==3.2.0
anyio==4.13.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.13.0
deprecation==2.1.0
feedparser==6.0.12
frozendict==2.4.7
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
msgpack==1.0.3
multidict==6.7.1
multitasking==0.0.12
numpy==2.4.0
packaging==26.0
pandas==2.3.3
peewee==4.0.3
platformdirs==4.9.4
propcache==0.4.1
protobuf==7.34.1
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
pytz==2025.2
PyYAML==6.0.1
requests==2.32.5
schedule==1.2.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
websocket-client==1.9.0
websockets==16.0
yarl==1.22.0
yfinance==1.2.0
